+++ graphql-maven-plugin-logic/src/test/java/com/graphql_java_generator/plugin/compilation_tests/AbstractIntegrationTest.java
@@ -62,6 +62,7 @@ abstract class AbstractIntegrationTest {
 		// Preparation
 		mavenTestHelper.deleteDirectoryAndContentIfExists(pluginConfiguration.getTargetSourceFolder());
 		mavenTestHelper.deleteDirectoryAndContentIfExists(pluginConfiguration.getTargetClassFolder());
+		pluginConfiguration.logConfiguration();
 
 		// Go, go, go
 		codeGenerator.generateCode();