+++ graphql-maven-plugin-samples/graphql-maven-plugin-samples-allGraphQLCases-pojo-client/src/test/java/org/forum/pojo/PojoTest.java
@@ -1,0 +1,11 @@
+package org.forum.pojo;
+
+import org.allGraphQLCases.client.pojo.AllFieldCases;
+
+class PojoTest {
+
+	private AllFieldCases allFieldCases;
+
+	// No test execution: the simple fact that this class compiles is enough.
+
+}